move some shared headers to -web
