+++ actix-http/src/header/shared/mod.rs
@@ -2,16 +2,12 @@
 
 mod charset;
 mod content_encoding;
-mod encoding;
-mod entity;
 mod extended;
-mod httpdate;
 mod quality_item;
+mod httpdate;
 
 pub use self::charset::Charset;
 pub use self::content_encoding::ContentEncoding;
-pub use self::encoding::Encoding;
-pub use self::entity::EntityTag;
 pub use self::extended::{parse_extended_value, ExtendedValue};
 pub use self::httpdate::HttpDate;
 pub use self::quality_item::{q, qitem, Quality, QualityItem};
